requirements.txt
fastapi[all]
torch
torchvision
requests
Pillow
redis
rq
python-multipart
mpld3